Moves NewsFeedView into its own new home
